# apps/backend/openapi.yaml
openapi: 3.1.0
info:
  title: DeRec reference app — backend
  version: "0.1.0"
  description: |
    The backend half of the DeRec reference app: an actor registry, a message
    relay, and a debug surface. It runs as a developer's local container and
    holds **no protocol logic** — every DeRec flow is executed by the SDK, in
    the browser or inside a provisioned actor.

    There is **no authentication**. That is deliberate: this is a debugging
    tool that exposes its own internals, and it is not meant to be reachable
    from anywhere but a developer's machine.

    ## Concepts

    - **Actor** — one participant. Either *browser-managed* (its protocol runs
      in a page; it receives by polling a mailbox here) or *provisioned* (its
      protocol runs in this process). Every actor advertises one or more
      **transports**; peers dial those.
    - **Helper pool** — provisioned helpers belong to the server, not to
      whoever asked for them. `POST /helpers/ensure` therefore states a
      *target*, and only the shortfall is created.
    - **Transport mode** — what a helper advertises: `http`, `grpc`, or
      `both`. Independent of what it can dial, which is always both.

    ## If you are an agent

    Start with `GET /debug/state` for everything that exists right now, and
    `GET /debug/events` for what has happened, in order, with the transport
    each message travelled over. Those two answer most questions without
...

...

    ## Errors

...

    | Status | When |
...
servers:
...
paths:
...
  /debug/events:
    get:
      tags: [debug]
      summary: What this server did, in order
      description: |
        A bounded window (2000 events) of every message this server delivered
        or refused, tagged with the transport that actually carried it.

        Poll by passing the previous response's `latest_seq` back as `after`.
        A non-zero `dropped` means the window lost its oldest entries, which is
        how you tell truncation from silence.
      parameters:
        - name: after
          in: query
          description: Return only events with a higher sequence number.
          schema: { type: integer, format: int64, default: 0 }
        - name: limit
          in: query
          description: Cap the page. Omitted or zero means the full window.
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: A page of the log
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EventSnapshot" }
...
  /derec/relay:
    post:
      tags: [transport]
      summary: Dial an endpoint on a browser owner's behalf
      description: |
        A browser cannot speak gRPC — there is no HTTP/2 trailer access from
        JavaScript. This backend already terminates transport for every actor,
        so it performs the dial instead, bounded by the same connect and
        request timeouts as its own calls.

        It delivers **only** to:

        1. **this node**, under any address it answers to now — its
           `base_url` host or `localhost`/`127.0.0.1`/`[::1]`, on the listen
           or public port — or advertised before (recorded across restarts, so
           a container republished on another port still answers to the old
           one). These are delivered in-process, never dialled: a `grpc://`
           target is routed by the envelope's channel id exactly as the gRPC
           listener routes it, an `http(s)://…/derec/<actor_id>` target goes to
           that actor's inbox;
        2. an endpoint a registered actor currently advertises;
        3. **another node** whose host or `host:port` the operator listed in
           `server.relay_allowed_hosts` (`DEREC_RELAY_ALLOWED_HOSTS`), or any
           node when that is `*`. Empty by default, so out of the box the relay
           reaches this node only; without the allowlist it would be an open
           proxy.

        `data` must decode to a DeRec envelope naming a channel, of at most
        4 MiB (what the gRPC listener accepts); anything else is refused before
        any delivery. Every outcome — refusals included, with their reason — is
        recorded in `/debug/events` with carrier `grpc_via_relay` or
        `http_via_relay`, and attributed to `actor_id` when the request names
        one.
      responses:
        "202": { description: Delivered }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403":
          description: No registered actor advertises that endpoint, and it is not this node's own gRPC listener
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
        "502":
          description: >
            The dial failed, or the target is this node and nothing here holds
            what the message names (an unknown actor or channel)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: >
            The relay is disabled (`grpc_relay_enabled = false`), or the
            recipient on this node has a full mailbox
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /derec/{actor_id}:
    post:
      tags: [transport]
      summary: Deliver a protocol message to an actor (HTTP transport)
      description: |
        The endpoint peers post to. The body is raw protobuf wire bytes, not
        JSON. A suspended actor answers `202` and discards the message, so a
        peer sees the same thing it would from a healthy one.

        A body may be at most 4 MiB — the most the gRPC listener accepts, so a
        message that travels one way travels every way.

        A browser actor's messages queue in the database until its tab polls.
        Each mailbox holds at most 1000 messages / 16 MiB of message bytes (as
        posted, not as stored); beyond that a new message is refused with
        `503` and nothing already queued is dropped.
      parameters:
        - $ref: "#/components/parameters/ActorId"
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema: { type: string, format: binary }
      responses:
        "202": { description: Accepted — delivered, or discarded because the actor is offline }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "500": { $ref: "#/components/responses/InternalError" }
        "503":
          description: The recipient's mailbox is full, or could not be written
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  parameters:
    ActorId:
      name: actor_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Malformed or contradictory request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: No such actor, channel, or contact
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    PayloadTooLarge:
      description: The request body is over the limit
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    UnsupportedMediaType:
      description: "Expected `Content-Type: application/json`"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    UnprocessableEntity:
      description: Well-formed JSON of the wrong shape
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    InternalError:
      description: This node failed; the detail is logged, not returned
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: A sentence a caller can act on. Never names an internal type.
...
    EventSnapshot:
      type: object
      properties:
        events:
          type: array
          items: { $ref: "#/components/schemas/Event" }
        dropped:
          type: integer
          description: How many events fell out of the retained window.
        latest_seq:
          type: integer
          description: Pass back as `after` to poll for what comes next.
    Event:
      type: object
      properties:
        seq: { type: integer }
        at_ms: { type: integer }
        direction: { type: string, enum: [inbound, outbound] }
        carrier:
          type: string
          enum: [http, grpc, grpc_via_relay, http_via_relay]
          description: |
            What actually carried this message — not what the peer advertises.
            `grpc_via_relay` and `http_via_relay` are a browser owner's message
            this backend dialled on its behalf.
        outcome: { type: string, enum: [delivered, dropped, refused] }
        actor_id:
          type: string
          format: uuid
          description: Absent when the message could not be routed, which is the interesting case.
        channel_id:
          type: string
          description: Absent when the message named no channel (or could not be decoded).
        bytes: { type: integer }
        detail: { type: string }